--- src/App.js
@@ -1,15 +1,0 @@
-const App = () => {
-  return (
-    <div className="App">
-      <div className="alert alert-success">Test Bootstrap</div>
-
-      <div className="test">
-        Test Bootstrap
-        <h1>Primary</h1>
-        <h3>Tertiary</h3>
-      </div>
-    </div>
-  );
-};
-
-export default App;
